fix: address code review findings
User-visible bugs:
- HistoryScreen Undo snackbar: SnackbarDuration.Short → Long. The comment
  promised "Long ≈ 10s" but the code passed Short (~4s), making the Undo
  affordance disappear ~1s before the VM's 5s commit window — silent loss
  of the recovery option.
- MeasureScreen showSaveDialog/showRationale: plain remember → rememberSaveable.
  Rotation during the save dialog wiped the dialog and the user's typed text
  along with it (the form fields had their own rememberSaveable, but the
  visibility flag did not).
- DetailViewModel.saveNote: differentiate validation failure from generic
  repository failure. Previously every onFailure showed "Note longer than
  200 characters" regardless of cause; Room IO errors or FK violations now
  surface a distinct "Couldn't save the note" message.
- DetailViewModel.init: wrap getMeasurementById in runCatching. A throw from
  the use-case (e.g. Room IO during cold start) previously pinned the screen
  at loading=true with no UI escape.
- SaveMeasurementUseCase: validate isFinite on avg/min/max dB. SQLite stores
  NaN/Infinity without complaint, which then poisons every UI surface
  formatting the row ("NaN dB" in History card, axis errors in Detail chart).

Test additions:
- SaveMeasurementUseCaseTest: 3 new cases (NaN avg, infinite min, infinite max).
- DetailViewModelTest: 2 new cases (init exception → not-found UX,
  saveNote repository failure → save_failed snackbar, not too_long).
- FakeMeasurementRepository: class promoted to `open` so tests can override
  individual operations to simulate Room failures inline.

Doc fixes:
- README: APK size 17.9MB → 28.2MB (Phase 3 baseline after Room+KSP+icons),
  Robolectric module list adds :core:data and :feature:history.
- core/testing/fakes/README.md: rewrote — old text claimed the directory
  was empty awaiting Phase 2.

Drive-by:
- Removed unused `kotlinx.coroutines.flow.collect` imports in HistoryScreen,
  MeasureScreen, DetailScreen (pre-existing detekt warnings).

+++ feature/history/src/main/kotlin/ru/dmdp/tishina/feature/history/HistoryScreen.kt
@@ -32,7 +32,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import kotlinx.coroutines.flow.collect
 import ru.dmdp.tishina.feature.history.ui.HistoryEmptyState
 import ru.dmdp.tishina.feature.history.ui.HistoryItemCard
 
@@ -69,12 +68,11 @@ fun HistoryScreen(
                     val result = snackbarHostState.showSnackbar(
                         message = context.getString(effect.messageRes),
                         actionLabel = context.getString(effect.actionRes),
-                        // Long ≈ 10 s — enough that even on slow-finger taps the user can
-                        // still hit Undo without the snackbar auto-dismissing. The VM
-                        // commits at 5 s regardless; if the snackbar lingers slightly
-                        // longer the action button just becomes a no-op (UndoConfirmed
-                        // with no pending delete is handled).
-                        duration = SnackbarDuration.Short,
+                        // Long ≈ 10 s — outlasts the VM's 5 s Undo window so the Undo
+                        // affordance is available for the whole window. After the VM
+                        // commits, UndoConfirmed becomes a safe no-op (handled by the
+                        // VM's "no pending delete" branch).
+                        duration = SnackbarDuration.Long,
                     )
                     if (result == SnackbarResult.ActionPerformed) {
                         viewModel.onEvent(HistoryUiEvent.UndoConfirmed)
